+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Unreleased Changes
   were being treated as valid next tokens for paging response. This appears
   to only have affected `Aws::DynamoDB::Client#batch_get_item`.
 
+* Issue - Instance Profile Credentials - Resolved an issue where instance
+  profile credentials would never load if the first attempt fails.
+  See GitHub issue #193.
+
 2.0.18 (2015-01-08)
 ------------------